- Linux/GCC (as usual :P) - Ported lockless code to gcc inline assembly - A few new net messages (gather, attack, add waypoint) - Support for new messages in network->entity order converter - Implemented rudimentary JS interface for Interaction - issueCommand JS API, connected to the network - Removed Interaction stuff now replaced by JS - And something in there should probably break VS builds :P
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@2316 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/Entity.h
@@ -106,7 +106,7 @@ public:
 	int m_lastState;
 	
 	// Position in the current state's cycle
-	static const size_t NOT_IN_CYCLE = -1;
+	static const size_t NOT_IN_CYCLE = (size_t)-1;
 	size_t m_fsm_cyclepos; // -cycle_length....cycle_length
 	CSkeletonAnim* m_fsm_animation; // the animation we're about to play this cycle,
 	size_t m_fsm_anipos; // the time at which we should start playing it.
@@ -197,6 +197,7 @@ public:
 	// Script-bound functions
 
 	jsval ToString( JSContext* cx, uintN argc, jsval* argv );
+	
 	bool Order( JSContext* cx, uintN argc, jsval* argv, bool Queued );
 	inline bool OrderSingle( JSContext* cx, uintN argc, jsval* argv )
 	{
